+++ src/main/java/com/ddudu/todo/domain/TodoStatus.java
@@ -1,4 +1,4 @@
-package com.ddudu.todo.model;
+package com.ddudu.todo.domain;
 
 public enum TodoStatus {
   UNCOMPLETED,
